requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.3
annotated-types==0.7.0
APScheduler==3.6.3
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
cachetools==4.2.2
certifi==2024.8.30
frozenlist==1.4.1
greenlet==3.0.3
idna==3.8
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
numpy==2.1.0
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.25
tornado==6.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
yarl==1.9.7
